adicionando emprestimo

+++ include/emprestimo.h
@@ -12,19 +12,19 @@ private:
     int id;
     long matricula;
     string data_emprestimo;
-    vector<int> codigo_livros_emprestado;
+    vector<long> codigo_livros_emprestado;
 
 public:
     Emprestimo(/* args */);
     ~Emprestimo();
-    Emprestimo(long matricula, const string &data_emprestimo, const std::vector<int> &codigo_livros_emprestado);
+    Emprestimo(long matricula, const string &data_emprestimo, const std::vector<long> &codigo_livros_emprestado);
     int getID() const;
     long getMatricula() const;
     string getDataEmprestimo() const;
-    vector<int> getCodigoLivrosEmprestado() const;
+    vector<long> getCodigoLivrosEmprestado() const;
 
     void setID(int id);
     void setMatricula(long matricula);
     void setDataEmprestimo(const string &data_emprestimo);
-    void setCodigoLivrosEmprestado(const vector<int>codigo_livros_emprestado);
+    void setCodigoLivrosEmprestado(const vector<long>codigo_livros_emprestado);
 };
